# Use JSMapIterator and JSSetIterator for deep equal comparisons