### What does this PR do?

This brings us back to parity in benchmarks for the fast path before commit e42dede. At the same time, we can now eliminate the slow path as the JSC iterators can work on classes that extend the native JSMap and JSSet types and achieve the same benchmarks in those cases.

I had the following results on an Apple M1 Pro:

This commit:

```
benchmark                time (avg)             (min … max)       p75       p99      p999
----------------------------------------------------------- -----------------------------
deepEqual Map           227 µs/iter       (218 µs … 281 µs)    232 µs    259 µs    270 µs
deepEqual CustomMap     228 µs/iter       (219 µs … 324 µs)    232 µs    264 µs    315 µs
```

```
benchmark                time (avg)             (min … max)       p75       p99      p999
----------------------------------------------------------- -----------------------------
deepEqual Set           155 µs/iter       (149 µs … 203 µs)    155 µs    186 µs    200 µs
deepEqual CustomSet     155 µs/iter       (149 µs … 347 µs)    151 µs    181 µs    191 µs
```

Before commit e42dede:

```
benchmark                time (avg)             (min … max)       p75       p99      p999
----------------------------------------------------------- -----------------------------
deepEqual Map           207 µs/iter       (200 µs … 324 µs)    204 µs    234 µs    293 µs
deepEqual CustomMap   2'316 ms/iter   (2'304 ms … 2'343 ms)  2'324 ms  2'343 ms  2'343 ms
```

```
benchmark                time (avg)             (min … max)       p75       p99      p999
----------------------------------------------------------- -----------------------------
deepEqual Set           130 µs/iter       (125 µs … 181 µs)    128 µs    158 µs    176 µs
deepEqual CustomSet   1'889 ms/iter   (1'885 ms … 1'899 ms)  1'895 ms  1'899 ms  1'899 ms
```

Fixes https://github.com/oven-sh/bun/issues/12498.

### How did you verify your code works?

- [x] I included a test for the new code, or existing tests cover it
- [x] I ran my tests locally and they pass (`bun-debug test test-file-name.test`)